requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.10
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow>=10.2.0
bleach==6.1.0
markdown==3.5.1
django-storages==1.14.2
boto3==1.29.7
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
drf-yasg==1.21.7
django-allauth==0.57.0
python-json-logger==2.0.7
